feat(bodies): bodies four times larger — up to 384x384 and 49152 cells
Slots come in three sizes (192 small, 48 medium, 16 large up to 384x384),
a body taking the smallest its box fits, so a body the size of a hall is
possible while the raster arenas take less memory than before. The
detector's region and cell limits and the detach ceiling are raised to
match, the fracture scratch indexes past sixteen bits, the substep ceiling
is raised so the largest body still cannot tunnel, and GameState and
Renderer are static rather than on main's stack.

+++ src/terrain_damage.c
@@ -208,7 +208,7 @@ static int TerrainDamageLabelComponents(TerrainDamageSystem *system,
         }
         ++count;
         system->component[index] = (uint8_t)count;
-        system->queue[tail++] = (uint16_t)index;
+        system->queue[tail++] = (uint32_t)index;
         while (head < tail) {
             static const int offsets[4][2] = {{0, -1}, {1, 0}, {0, 1}, {-1, 0}};
             int current = (int)system->queue[head++];
@@ -235,7 +235,7 @@ static int TerrainDamageLabelComponents(TerrainDamageSystem *system,
                     continue;
                 }
                 system->component[neighbour] = (uint8_t)count;
-                system->queue[tail++] = (uint16_t)neighbour;
+                system->queue[tail++] = (uint32_t)neighbour;
             }
         }
         sizes[count - 1] = size;
